Move src/css to src/core/dom (and test)
